make it so that float can be entered in second number

+++ calculator.py
@@ -30,7 +30,7 @@ class Calculator(CalculatorColor):
         while True:
             try:
                 self.first_num = float(input("Please input your First Number: "))
-                self.second_num = int(input("Please input your Second Number: "))
+                self.second_num = float(input("Please input your Second Number: "))
                 break
             except ValueError:
                 print(f"{self.error_color}Input numbers only{self.reset_color}")
